+++ CHANGELOG.md
@@ -11,6 +11,30 @@ Each version will have a separate `Breaking Changes` section as well. To describ
 
 ### Added 🎉
 
+* feat: new BLS aggregation service interface by @maximopalopoli in <https://github.com/Layr-Labs/eigensdk-go/pull/578>
+  * The new interface implies starting the service before using it, interact with it using a handler and receiving the aggregated responses in a separate channel.
+  * An example using the interface is:
+
+    ```Go
+    // initialize service
+    blsAgg := NewBlsAggregatorBuilder(fakeAvsRegistryService, hashFunction, logger)
+    handler, aggResponsesC := blsAgg.Start()
+
+    // Initialize task
+    metadata := NewTaskMetadata(taskIndex, blockNum, quorumNumbers, quorumThresholdPercentages, tasksTimeToExpiry)
+    err := handler.InitializeNewTask(metadata)
+
+    // Process signature
+    taskSignature := NewTaskSignature(taskIndex, taskResponse, blsSig, testOperator1.OperatorId)
+    err = handler.ProcessNewSignature(
+      context.Background(),
+      taskSignature,
+    )
+
+    // Receive responses
+    aggregationServiceResponse := <-aggResponsesC
+    ```
+
 * Added field `DontUseAllocationManager` to `BuildAllConfig` in [#580](https://github.com/Layr-Labs/eigensdk-go/pull/580)
 
 ### Changed